requirements.txt
# Core Flask dependencies
flask==2.2.3
flask-login
flask-sqlalchemy
flask-wtf==1.1.1
werkzeug==2.2.3

# Database
psycopg2-binary
sqlalchemy

# Document processing
python-docx
docxtpl==0.16.7

# Image processing
pillow>=10.4.0

# HTTP requests and web scraping
requests
beautifulsoup4
lxml

# Utilities
python-dotenv==1.0.0
itsdangerous

# Production server
gunicorn
greenlet

# Windows compatibility (conditional)
pywin32>=307; sys_platform == 'win32'

# Security and session management
cryptography
flask-session

# Database migrations
flask-migrate
alembic

# Task scheduling
schedule

# System monitoring
psutil

# Configuration management
pyyaml
watchdog
marshmallow

# Secrets management
hvac
requests

# Caching and performance
redis<4.0.0
celery>=5.2.0
flower>=1.2.0     # Celery monitoring
redis-py-cluster>=2.1.0

# Testing dependencies
pytest>=6.0.0
pytest-cov>=2.10.0
pytest-mock>=3.6.0
pytest-xdist>=2.2.0
pytest-html>=3.1.0
pytest-asyncio>=0.18.0
factory-boy>=3.2.0
faker>=13.0.0
coverage>=6.0.0
reportlab>=3.6.0  # For PDF testing
Pillow>=8.0.0     # For image testing
boto3
celery
click
flask-login
flask-session
flask-sqlalchemy
flask-wtf
marshmallow
python-dotenv
PyYAML
redis
watchdog
flask-migrate
flask-restx
psutil
PyJWT
schedule
pyotp
qrcode[pil]
bleach
structlog
waitress